Factor out handling of the remaining messages.
